🚀 Release Final: Arquitectura V2, WebSockets, ALNS+HGA, y Corrección de Tiempo

+++ backend/src/main/java/com/tasfb2b/planificador/web/SimulationStatusDTO.java
@@ -83,7 +83,7 @@ public class SimulationStatusDTO {
     /** Métricas de colapso */
     private Boolean isCollapseMode;
     private Integer rescuedFlights;
-    private Integer stressFactor;
+    private Double stressFactor;
     private Long startEpoch;
     private String algorithm;
